att_gt_het2

att_gt_het2 computes the difference of average treatment effects across two subpopulations in DID setups where there are more than two periods of data and allowing for treatment to occur at different points in time. Here, we assume same trends between the two supopulations. See Marcus ans Sant'Anna (2020) for a detailed description.

att_gt_het2(
  outcome,
  data,
  tname,
  idname = NULL,
  first.treat.name,
  nevertreated = T,
  het,
  aggte = TRUE,
  maxe = NULL,
  mine = NULL,
  w = NULL,
  alp = 0.05,
  bstrap = T,
  biters = 1000,
  clustervars = NULL,
  cband = T,
  printdetails = TRUE,
  seedvec = NULL,
  pl = FALSE,
  cores = 2,
  method = "logit"
)

Arguments

outcome
The outcome y (in quotations, always!)

data
The name of the data.frame that contains the data

tname
The name of the column containing the time periods

idname
The individual (cross-sectional unit) id name

first.treat.name
The name of the variable in data that contains the first period when a particular observation is treated. This should be a positive number for all observations in treated groups. It should be 0 for observations in the untreated group.

nevertreated
Boolean for using the group which is never treated in the sample as the comparison unit. Default is TRUE.

het
The name of the column containing the (binary) categories for heterogeneity

aggte
boolean for whether or not to compute aggregate treatment effect parameters, default TRUE

maxe
maximum values of periods ahead to be computed in event study. Only used if aggte = T.

mine
minimum values of periods ahead to be computed in event study. Only used if aggte = T.

w
The name of the column containing the sampling weights. If not set, all observations have same weight.

alp
the significance level, default is 0.05

bstrap
Boolean for whether or not to compute standard errors using the multiplier boostrap. If standard errors are clustered, then one must set bstrap=TRUE. Default is TRUE.

biters
The number of boostrap iterations to use. The default is 1000, and this is only applicable if bstrap=TRUE.

clustervars
A vector of variables to cluster on. At most, there can be two variables (otherwise will throw an error) and one of these must be the same as idname which allows for clustering at the individual level.

cband
Boolean for whether or not to compute a uniform confidence band that covers all of the group-time average treatment effects with fixed probability 1-alp. The default is TRUE and the resulting standard errors will be pointwise.

printdetails
Boolean for showing detailed results or not

seedvec
Optional value to set random seed; can possibly be used in conjunction with bootstrapping standard errors#' (not implemented)

pl
Boolean for whether or not to use parallel processing (not implemented) is TRUE.

cores
The number of cores to use for parallel processing (not implemented)

method
The method for estimating the propensity score when covariates are included (not implemented)

Value

MP object

References

Callaway, Brantly and Sant'Anna, Pedro. "Difference-in-Differences with Multiple Time Periods and an Application on the Minimum Wage and Employment." Working Paper <https://ssrn.com/abstract=3148250> (2018).