compute.att_gt_het2

compute.att_gt_het does the main work for computing mutliperiod group-time average treatment effects with heterogeneity

compute.att_gt_het2(
  flen,
  tlen,
  flist,
  tlist,
  data,
  dta,
  first.treat.name,
  outcome,
  tname,
  idname,
  method,
  seedvec,
  pl,
  cores,
  printdetails,
  nevertreated,
  het
)

Arguments

data
The name of the data.frame that contains the data

first.treat.name
The name of the variable in data that contains the first period when a particular observation is treated. This should be a positive number for all observations in treated groups. It should be 0 for observations in the untreated group.

outcome
The outcome y (in quotations, always!)

tname
The name of the column containing the time periods

idname
The individual (cross-sectional unit) id name

method
The method for estimating the propensity score when covariates are included (not implemented)

seedvec
Optional value to set random seed; can possibly be used in conjunction with bootstrapping standard errors#' (not implemented)

pl
Boolean for whether or not to use parallel processing (not implemented) is TRUE.

cores
The number of cores to use for parallel processing (not implemented)

printdetails
Boolean for showing detailed results or not

nevertreated
Boolean for using the group which is never treated in the sample as the comparison unit. Default is TRUE.

het
The name of the column containing the (binary) categories for heterogeneity

Value

a list with length equal to the number of groups times the number of time periods; each element of the list contains an object that contains group-time average treamtent effect as well as which group it is for and which time period it is for. It also exports the influence function which is used externally to compute standard errors.