gplot

does the heavy lifting for making a plot of an group-time average treatment effect

gplot(ssresults, ylim = NULL, xlab = NULL, ylab = NULL,
  title = "Group", xgap = 1)

Arguments

ylim
optional y limits for the plot; settng here makes the y limits the same across different plots

xlab
optional x-axis label

ylab
optional y-axis label

title
optional plot title

xgap
optional gap between the labels on the x-axis. For example, xgap=3 indicates that the labels should show up for every third value on the x-axis. The default is 1.

Value

a ggplot2 object