AGGTE

AGGTE class for aggregate treatment effects

AGGTE(simple.att = NULL, simple.se = NULL, selective.att = NULL,
  selective.se = NULL, selective.att.g = NULL, selective.se.g = NULL,
  dynamic.att = NULL, dynamic.se = NULL, dynamic.att.e = NULL,
  dynamic.se.e = NULL, calendar.att = NULL, calendar.se = NULL,
  calendar.att.t = NULL, calendar.se.t = NULL, dynsel.att.e1 = NULL,
  dynsel.se.e1 = NULL, dynsel.att.ee1 = NULL, dynsel.se.ee1 = NULL)

Arguments

simple.att
simple weighted average of group-time average treatment effects

simple.se
the standard error for simple.att

selective.att
aggregated group-time average treament effects when there is selective treatment timing

selective.se
the standard error for selective.att

selective.att.g
aggregated group-time average treatment effects when there is selective treatment timing for each particular group

selective.se.g
the standard error for selective.att.g

dynamic.att
aggregated group-time average treatment effects when there are dynamic treatment effects

dynamic.se
the standard error for dynamic.att

dynamic.att.e
aggregated group-time average treatment effects when there are dynamic treatment effects for each length of exposure to treatment

dynamic.se.e
the standard error for dynamic.att.e

calendar.att
the aggregated group-time average treatment effects when there are calendar time effects

calendar.se
the standard error for calendar.att

calendar.att.t
the aggregated group-time average treatment effects when there are calendar time effects for each time period

calendar.se.t
the standard error for calendar.att.t

dynsel.att.e1
aggregated group-time average treatment effects when there are dynamic treatment effects and selective treatment timing. Here, e1 is the number of periods that a group is required to be treated in order to be included in the results.

dynsel.se.e1
the standard error for dynsel.att.e1

dynsel.att.ee1
aggregated group-time average treatment effects when there are dynamic treatment effects and selective treatment timing. Here, e1 is the number of periods that a group is required to be treated in order to be included in the results and for each length of exposure to treatment

dynsel.se.ee1
the standard error for dynsel.att.ee1