Group-Time Average Treatment Effects

att_gt computes average treatment effects in DID setups where there are more than two periods of data and allowing for treatment to occur at different points in time and allowing for treatment effect heterogeneity and dynamics. See Callaway and Sant'Anna (2020) for a detailed description.

att_gt(
  yname,
  tname,
  idname = NULL,
  gname,
  xformla = NULL,
  data,
  panel = TRUE,
  allow_unbalanced_panel = FALSE,
  control_group = c("nevertreated", "notyettreated"),
  anticipation = 0,
  weightsname = NULL,
  alp = 0.05,
  bstrap = TRUE,
  cband = TRUE,
  biters = 1000,
  clustervars = NULL,
  est_method = "dr",
  print_details = FALSE,
  pl = FALSE,
  cores = 1
)

Arguments

yname
The name of the outcome variable

tname
The name of the column containing the time periods

idname
The individual (cross-sectional unit) id name

gname
The name of the variable in data that contains the first period when a particular observation is treated. This should be a positive number for all observations in treated groups. It defines which "group" a unit belongs to. It should be 0 for units in the untreated group.

xformla
A formula for the covariates to include in the model. It should be of the form ~ X1 + X2. Default is NULL which is equivalent to xformla=~1. This is used to create a matrix of covariates which is then passed to the 2x2 DID estimator chosen in est_method.

data
The name of the data.frame that contains the data

panel
Whether or not the data is a panel dataset. The panel dataset should be provided in long format -- that is, where each row corresponds to a unit observed at a particular point in time. The default is TRUE. When is using a panel dataset, the variable idname must be set. When panel=FALSE, the data is treated as repeated cross sections.

allow_unbalanced_panel
Whether or not function should "balance" the panel with respect to time and id. The default values if FALSE which means that att_gt will drop all observations where data is not observed in all periods. The advantage of this is that the computations are faster (sometimes substantially).

control_group
Which units to use the control group. The default is "nevertreated" which sets the control group to be the group of units that never participate in the treatment. This group does not change across groups or time periods. The other option is to set group="notyettreated". In this case, the control group is set to the group of units that have not yet participated in the treatment in that time period. This includes all never treated units, but it includes additional units that eventually participate in the treatment, but have not participated yet.

anticipation
The number of time periods before participating in the treatment where units can anticipate participating in the treatment and therefore it can affect their untreated potential outcomes

weightsname
The name of the column containing the sampling weights. If not set, all observations have same weight.

alp
the significance level, default is 0.05

bstrap
Boolean for whether or not to compute standard errors using the multiplier boostrap. If standard errors are clustered, then one must set bstrap=TRUE. Default is TRUE (in addition, cband is also by default TRUE indicating that uniform confidence bands will be returned. If bstrap is FALSE, then analytical standard errors are reported.

cband
Boolean for whether or not to compute a uniform confidence band that covers all of the group-time average treatment effects with fixed probability 1-alp. In order to compute uniform confidence bands, bstrap must also be set to TRUE. The default is TRUE.

biters
The number of boostrap iterations to use. The default is 1000, and this is only applicable if bstrap=TRUE.

clustervars
A vector of variables to cluster on. At most, there can be two variables (otherwise will throw an error) and one of these must be the same as idname which allows for clustering at the individual level.

est_method
the method to compute group-time average treatment effects. The default is "dr" which uses the doubly robust approach in the DRDID package. Other built-in methods include "ipw" for inverse probability weighting and "reg" for first step regression estimators. The user can also pass their own function for estimating group time average treatment effects. This should be a function f(Y1,Y0,treat,covariates) where Y1 is an n x 1 vector of outcomes in the post-treatment outcomes, Y0 is an n x 1 vector of pre-treatment outcomes, treat is a vector indicating whether or not an individual participates in the treatment, and covariates is an n x k matrix of covariates. The function should return a list that includes ATT (an estimated average treatment effect), and inf.func (an n x 1 influence function). The function can return other things as well, but these are the only two that are required. est_method is only used if covariates are included.

print_details
Whether or not to show details/progress of computations. Default is FALSE.

pl
Whether or not to use parallel processing (not implemented yet)

cores
The number of cores to use for parallel processing (not implemented yet)

Value

an MP object containing all the results for group-time average treatment effects

References

Callaway, Brantly and Sant'Anna, Pedro H. C. "Difference-in-Differences with Multiple Time Periods" Forthcoming at the Journal of Econometrics <https://arxiv.org/abs/1803.09015> (2020).

Examples

data(mpdta)
search()

#>  [1] ".GlobalEnv"        "package:did"       "package:stats"    
#>  [4] "package:graphics"  "package:grDevices" "package:utils"    
#>  [7] "package:datasets"  "package:methods"   "Autoloads"        
#> [10] "tools:callr"       "package:base"     

lpa <- .libPaths()
lpa

#> [1] "/private/var/folders/yh/wqt796_d5vg2bdyz_sbr66zw0000gq/T/RtmpkcpgQw/temp_libpath165e13ceb2b05"
#> [2] "/Library/Frameworks/R.framework/Versions/4.0/Resources/library"                               

sapply(lpa, function(lp) list.files(lp))

#> $`/private/var/folders/yh/wqt796_d5vg2bdyz_sbr66zw0000gq/T/RtmpkcpgQw/temp_libpath165e13ceb2b05`
#> [1] "did"
#> 
#> $`/Library/Frameworks/R.framework/Versions/4.0/Resources/library`
#>   [1] "abind"                "acepack"              "AER"                 
#>   [4] "alphavantager"        "animation"            "anytime"             
#>   [7] "ape"                  "argon2"               "arm"                 
#>  [10] "askpass"              "assertthat"           "backports"           
#>  [13] "bacondecomp"          "badger"               "base"                
#>  [16] "base64enc"            "BBmisc"               "bbotk"               
#>  [19] "bdsmatrix"            "BH"                   "bibtex"              
#>  [22] "BiocManager"          "BiocVersion"          "bitops"              
#>  [25] "blob"                 "BMisc"                "bookdown"            
#>  [28] "boot"                 "boot.heterogeneity"   "bootcluster"         
#>  [31] "brew"                 "brio"                 "broom"               
#>  [34] "callr"                "car"                  "carData"             
#>  [37] "caTools"              "CausalGAM"            "CBPS"                
#>  [40] "cellranger"           "checkmate"            "class"               
#>  [43] "cli"                  "clipr"                "cluster"             
#>  [46] "clusterGeneration"    "clusterPower"         "coda"                
#>  [49] "codetools"            "colorspace"           "combinat"            
#>  [52] "commonmark"           "compiler"             "conflicted"          
#>  [55] "conquer"              "corrplot"             "covr"                
#>  [58] "cowplot"              "cpp11"                "crayon"              
#>  [61] "crosstalk"            "csabounds"            "cubature"            
#>  [64] "curl"                 "cvAUC"                "data.table"          
#>  [67] "datasets"             "DBI"                  "dbplyr"              
#>  [70] "DDD"                  "delayed"              "DEoptimR"            
#>  [73] "Deriv"                "desc"                 "deSolve"             
#>  [76] "devtools"             "dfidx"                "did"                 
#>  [79] "diffobj"              "digest"               "diptest"             
#>  [82] "dlstats"              "doBy"                 "doParallel"          
#>  [85] "doRNG"                "doSNOW"               "downlit"             
#>  [88] "dplyr"                "DRDID"                "DT"                  
#>  [91] "dygraphs"             "elasso"               "ellipsis"            
#>  [94] "evaluate"             "eventstudies"         "EventStudy"          
#>  [97] "expm"                 "fansi"                "farver"              
#> [100] "fastDummies"          "fastmap"              "fastmatch"           
#> [103] "fastplm"              "fda"                  "fect"                
#> [106] "flexclust"            "flexmix"              "forcats"             
#> [109] "foreach"              "forecast"             "foreign"             
#> [112] "Formula"              "formula.tools"        "fpc"                 
#> [115] "fracdiff"             "fs"                   "furrr"               
#> [118] "future"               "future.apply"         "gam"                 
#> [121] "gbRd"                 "gdata"                "generics"            
#> [124] "getPass"              "GGally"               "ggforce"             
#> [127] "ggplot2"              "ggpubr"               "ggrepel"             
#> [130] "ggsci"                "ggsignif"             "ggthemes"            
#> [133] "gh"                   "git2r"                "glmnet"              
#> [136] "glmx"                 "globals"              "glue"                
#> [139] "gower"                "gplots"               "graphics"            
#> [142] "grDevices"            "grid"                 "gridExtra"           
#> [145] "gtable"               "gtools"               "hal9001"             
#> [148] "harvestr"             "haven"                "here"                
#> [151] "hettreatreg"          "hexbin"               "highlight"           
#> [154] "highr"                "Hmisc"                "hms"                 
#> [157] "htmlTable"            "htmltools"            "htmlwidgets"         
#> [160] "httpuv"               "httr"                 "hunspell"            
#> [163] "icon"                 "igraph"               "imputeMissings"      
#> [166] "ini"                  "ipred"                "IPS"                 
#> [169] "isoband"              "iterators"            "ivpack"              
#> [172] "janitor"              "jpeg"                 "jsonlite"            
#> [175] "kableExtra"           "kernlab"              "KernSmooth"          
#> [178] "kmte"                 "knitcitations"        "knitLatex"           
#> [181] "knitr"                "knitrBootstrap"       "knitrProgressBar"    
#> [184] "labeling"             "LassoNet"             "later"               
#> [187] "lattice"              "latticeExtra"         "lava"                
#> [190] "lazyeval"             "lemon"                "lfe"                 
#> [193] "lgr"                  "lifecycle"            "listenv"             
#> [196] "lme4"                 "lmtest"               "longbowtools"        
#> [199] "lspartition"          "ltmle"                "lubridate"           
#> [202] "magick"               "magrittr"             "manipulateWidget"    
#> [205] "mapdata"              "maps"                 "maptools"            
#> [208] "markdown"             "MASS"                 "Matching"            
#> [211] "MatchIt"              "matlib"               "Matrix"              
#> [214] "matrixcalc"           "MatrixModels"         "matrixStats"         
#> [217] "maxLik"               "mclust"               "memoise"             
#> [220] "metafor"              "methods"              "mgcv"                
#> [223] "mi"                   "mime"                 "miniUI"              
#> [226] "minqa"                "miscTools"            "mlbench"             
#> [229] "mlogit"               "mlr3"                 "mlr3data"            
#> [232] "mlr3learners"         "mlr3measures"         "mlr3misc"            
#> [235] "mlr3tuning"           "mnormt"               "modeldata"           
#> [238] "modelr"               "modeltools"           "msm"                 
#> [241] "munsell"              "mvtnorm"              "nlme"                
#> [244] "nloptr"               "nnet"                 "nnls"                
#> [247] "np"                   "numDeriv"             "openssl"             
#> [250] "openxlsx"             "operator.tools"       "origami"             
#> [253] "packrat"              "pacman"               "padr"                
#> [256] "pander"               "panelView"            "paradox"             
#> [259] "parallel"             "patchwork"            "pbapply"             
#> [262] "pbdZMQ"               "pbkrtest"             "pbmcapply"           
#> [265] "PerformanceAnalytics" "phangorn"             "phytools"            
#> [268] "pillar"               "pkgbuild"             "pkgconfig"           
#> [271] "pkgdown"              "pkgload"              "plm"                 
#> [274] "plogr"                "plotly"               "plotrix"             
#> [277] "plyr"                 "png"                  "polyclip"            
#> [280] "polynom"              "prabclus"             "pracma"              
#> [283] "praise"               "prettyunits"          "processx"            
#> [286] "prodlim"              "profvis"              "progress"            
#> [289] "promises"             "PRROC"                "ps"                  
#> [292] "pstest"               "purrr"                "qte"                 
#> [295] "quadprog"             "Quandl"               "quantmod"            
#> [298] "quantreg"             "quantreg.nonpar"      "R.methodsS3"         
#> [301] "R.oo"                 "R6"                   "ragg"                
#> [304] "randomForest"         "raster"               "rbibutils"           
#> [307] "rcmdcheck"            "RColorBrewer"         "Rcpp"                
#> [310] "RcppArmadillo"        "RcppEigen"            "Rdpack"              
#> [313] "readr"                "readstata13"          "readxl"              
#> [316] "Rearrangement"        "recipes"              "RefManageR"          
#> [319] "rematch"              "rematch2"             "remote"              
#> [322] "remoter"              "remotes"              "repr"                
#> [325] "reprex"               "reshape"              "reshape2"            
#> [328] "rex"                  "rgl"                  "riingo"              
#> [331] "rio"                  "rJava"                "RJSONIO"             
#> [334] "rlang"                "rmarkdown"            "rmdexamples"         
#> [337] "rngtools"             "robustbase"           "ROCR"                
#> [340] "roxygen2"             "rpart"                "rprojroot"           
#> [343] "rsample"              "rsconnect"            "Rsolnp"              
#> [346] "rstackdeque"          "rstatix"              "rstudioapi"          
#> [349] "rvcheck"              "rversions"            "rvest"               
#> [352] "sandwich"             "scales"               "scatterplot3d"       
#> [355] "selectr"              "sem"                  "sessioninfo"         
#> [358] "sets"                 "shape"                "shiny"               
#> [361] "SiZer"                "skimr"                "sl3"                 
#> [364] "slider"               "snakecase"            "snow"                
#> [367] "sourcetools"          "sp"                   "SparseM"             
#> [370] "spatial"              "speedglm"             "spelling"            
#> [373] "splines"              "SQUAREM"              "stargazer"           
#> [376] "statmod"              "stats"                "stats4"              
#> [379] "stringi"              "stringr"              "subplex"             
#> [382] "SuperLearner"         "survival"             "sys"                 
#> [385] "systemfonts"          "tcltk"                "testit"              
#> [388] "testthat"             "texreg"               "textshaping"         
#> [391] "tibble"               "tidyfast"             "tidyquant"           
#> [394] "tidyr"                "tidyselect"           "tidyverse"           
#> [397] "timeDate"             "timetk"               "tinytex"             
#> [400] "tlverse"              "tmle"                 "tmle3"               
#> [403] "tmle3mopttx"          "tmle3shift"           "tmvnsim"             
#> [406] "tools"                "translations"         "truncnorm"           
#> [409] "trust"                "tseries"              "TTR"                 
#> [412] "tufte"                "tweenr"               "urca"                
#> [415] "usethis"              "utf8"                 "utils"               
#> [418] "uuid"                 "vctrs"                "viridis"             
#> [421] "viridisLite"          "visNetwork"           "waldo"               
#> [424] "warp"                 "webshot"              "whisker"             
#> [427] "withr"                "xfun"                 "xlsx"                
#> [430] "xlsxjars"             "xml2"                 "xopen"               
#> [433] "xtable"               "xts"                  "yaml"                
#> [436] "zip"                  "zoo"                 
#> 

# with covariates
out1 <- att_gt(yname="lemp",
               tname="year",
               idname="countyreal",
               gname="first.treat",
               xformla=~lpop,
               data=mpdta)
summary(out1)

#> 
#> Call:
#> att_gt(yname = "lemp", tname = "year", idname = "countyreal", 
#>     gname = "first.treat", xformla = ~lpop, data = mpdta)
#> 
#> Reference: Callaway, Brantly and Pedro H.C. Sant'Anna.  "Difference-in-Differences with Multiple Time Periods." Forthcoming at the Journal of Econometrics <https://arxiv.org/abs/1803.09015>, 2020. 
#> 
#> Group-Time Average Treatment Effects:
#>  Group Time ATT(g,t) Std. Error [95% Simult.  Conf. Band]  
#>   2004 2004  -0.0145     0.0234       -0.0761      0.0470  
#>   2004 2005  -0.0764     0.0296       -0.1541      0.0013  
#>   2004 2006  -0.1404     0.0396       -0.2445     -0.0364 *
#>   2004 2007  -0.1069     0.0329       -0.1934     -0.0204 *
#>   2006 2004  -0.0005     0.0217       -0.0575      0.0565  
#>   2006 2005  -0.0062     0.0189       -0.0557      0.0433  
#>   2006 2006   0.0010     0.0192       -0.0495      0.0515  
#>   2006 2007  -0.0413     0.0185       -0.0898      0.0072  
#>   2007 2004   0.0267     0.0143       -0.0108      0.0642  
#>   2007 2005  -0.0046     0.0159       -0.0463      0.0371  
#>   2007 2006  -0.0284     0.0196       -0.0799      0.0230  
#>   2007 2007  -0.0288     0.0161       -0.0710      0.0134  
#> ---
#> Signif. codes: `*' confidence band does not cover 0
#> 
#> P-value for pre-test of parallel trends assumption:  0.2816
#> Control Group:  Never Treated,  Anticipation Periods:  0
#> Estimation Method:  Doubly Robust

# without covariates
out2 <- att_gt(yname="lemp",
               tname="year",
               idname="countyreal",
               gname="first.treat",
               xformla=NULL,
               data=mpdta)
summary(out2)

#> 
#> Call:
#> att_gt(yname = "lemp", tname = "year", idname = "countyreal", 
#>     gname = "first.treat", xformla = NULL, data = mpdta)
#> 
#> Reference: Callaway, Brantly and Pedro H.C. Sant'Anna.  "Difference-in-Differences with Multiple Time Periods." Forthcoming at the Journal of Econometrics <https://arxiv.org/abs/1803.09015>, 2020. 
#> 
#> Group-Time Average Treatment Effects:
#>  Group Time ATT(g,t) Std. Error [95% Simult.  Conf. Band]  
#>   2004 2004  -0.0105     0.0251       -0.0770      0.0560  
#>   2004 2005  -0.0704     0.0324       -0.1562      0.0154  
#>   2004 2006  -0.1373     0.0401       -0.2436     -0.0309 *
#>   2004 2007  -0.1008     0.0366       -0.1977     -0.0039 *
#>   2006 2004   0.0065     0.0254       -0.0606      0.0737  
#>   2006 2005  -0.0028     0.0199       -0.0555      0.0499  
#>   2006 2006  -0.0046     0.0190       -0.0551      0.0459  
#>   2006 2007  -0.0412     0.0219       -0.0992      0.0168  
#>   2007 2004   0.0305     0.0158       -0.0113      0.0724  
#>   2007 2005  -0.0027     0.0165       -0.0465      0.0411  
#>   2007 2006  -0.0311     0.0179       -0.0784      0.0162  
#>   2007 2007  -0.0261     0.0163       -0.0692      0.0171  
#> ---
#> Signif. codes: `*' confidence band does not cover 0
#> 
#> P-value for pre-test of parallel trends assumption:  0.19331
#> Control Group:  Never Treated,  Anticipation Periods:  0
#> Estimation Method:  Doubly Robust