Group-Time Average Treatment Effects

att_gt computes average treatment effects in DID setups where there are more than two periods of data and allowing for treatment to occur at different points in time and allowing for treatment effect heterogeneity and dynamics. See Callaway and Sant'Anna (2019) for a detailed description.

att_gt(
  yname,
  tname,
  idname = NULL,
  first.treat.name,
  xformla = NULL,
  data,
  panel = TRUE,
  control.group = c("nevertreated", "notyettreated"),
  weightsname = NULL,
  alp = 0.05,
  bstrap = FALSE,
  cband = FALSE,
  biters = 1000,
  clustervars = NULL,
  estMethod = "dr",
  printdetails = TRUE,
  pl = FALSE,
  cores = 1
)

Arguments

yname
The name of the outcome variable

tname
The name of the column containing the time periods

idname
The individual (cross-sectional unit) id name

first.treat.name
The name of the variable in data that contains the first period when a particular observation is treated. This should be a positive number for all observations in treated groups. It defines which "group" a unit belongs to. It should be 0 for units in the untreated group.

xformla
A formula for the covariates to include in the model. It should be of the form ~ X1 + X2. Default is NULL which is equivalent to xformla=~1. This is used to create a matrix of covariates which is then passed to the 2x2 DID estimator chosen in estMethod.

data
The name of the data.frame that contains the data

panel
Whether or not the data is a panel dataset. The panel dataset should be provided in long format -- that is, where each row corresponds to a unit observed at a particular point in time. The default is TRUE. When is using a panel dataset, the variable idname must be set. When panel=FALSE, the data is treated as repeated cross sections.

control.group
Which units to use the control group. The default is "nevertreated" which sets the control group to be the group of units that never participate in the treatment. This group does not change across groups or time periods. The other option is to set group="notyettreated". In this case, the control group is set to the group of units that have not yet participated in the treatment in that time period. This includes all never treated units, but it includes additional units that eventually participate in the treatment, but have not participated yet.

weightsname
The name of the column containing the sampling weights. If not set, all observations have same weight.

alp
the significance level, default is 0.05

bstrap
Boolean for whether or not to compute standard errors using the multiplier boostrap. If standard errors are clustered, then one must set bstrap=TRUE. Default is FALSE and analytical standard errors are reported.

cband
Boolean for whether or not to compute a uniform confidence band that covers all of the group-time average treatment effects with fixed probability 1-alp. In order to compute uniform confidence bands, bstrap must also be set to TRUE. The default is FALSE and the resulting standard errors will be pointwise.

biters
The number of boostrap iterations to use. The default is 1000, and this is only applicable if bstrap=TRUE.

clustervars
A vector of variables to cluster on. At most, there can be two variables (otherwise will throw an error) and one of these must be the same as idname which allows for clustering at the individual level.

estMethod
the method to compute group-time average treatment effects. The default is "dr" which uses the doubly robust approach in the DRDID package. Other built-in methods include "ipw" for inverse probability weighting and "reg" for first step regression estimators. The user can also pass their own function for estimating group time average treatment effects. This should be a function f(Y1,Y0,treat,covariates) where Y1 is an n x 1 vector of outcomes in the post-treatment outcomes, Y0 is an n x 1 vector of pre-treatment outcomes, treat is a vector indicating whether or not an individual participates in the treatment, and covariates is an n x k matrix of covariates. The function should return a list that includes ATT (an estimated average treatment effect), and inf.func (an n x 1 influence function). The function can return other things as well, but these are the only two that are required. estMethod is only used if covariates are included.

printdetails
Boolean for showing detailed results or not

pl
Boolean for whether or not to use parallel processing (not implemented yet)

cores
The number of cores to use for parallel processing (not implemented yet)

nevertreated
did currently provides two possibilities for the control group. The first one is to use the group of units that are never participate in the treatment. This is the default behavior and occurs when nevertreated=TRUE. In the case where there is no group that is never treated, att_gt will use the group that is treated in the last period and provide a warning. The other possible control group is the "not yet treated" group. This is a larger group and includes both never treated units and units that have not yet (but eventually will) participate in the treatment. This control group changes across different time periods.

Value

MP object

References

Callaway, Brantly and Sant'Anna, Pedro H. C.. "Difference-in-Differences with Multiple Time Periods and an Application on the Minimum Wage and Employment." Working Paper <https://ssrn.com/abstract=3148250> (2019).

Examples

data(mpdta)

# with covariates
out1 <- att_gt(yname="lemp",
               tname="year",
               idname="countyreal",
               first.treat.name="first.treat",
               xformla=~lpop,
               data=mpdta,
               printdetails=FALSE)
summary(out1)

#> 
#> Reference: Callaway, Brantly and Sant'Anna, Pedro.  "Difference-in-Differences with Multiple Time Periods." Working Paper <https://ssrn.com/abstract=3148250>, 2019. 
#> 
#> 
#> 
#> | group| time|        att|        se|
#> |-----:|----:|----------:|---------:|
#> |  2004| 2004| -0.0145297| 0.0221292|
#> |  2004| 2005| -0.0764219| 0.0286713|
#> |  2004| 2006| -0.1404483| 0.0353781|
#> |  2004| 2007| -0.1069039| 0.0328865|
#> |  2006| 2004| -0.0004721| 0.0222234|
#> |  2006| 2005| -0.0062025| 0.0184957|
#> |  2006| 2006|  0.0009606| 0.0194002|
#> |  2006| 2007| -0.0412939| 0.0197211|
#> |  2007| 2004|  0.0267278| 0.0140657|
#> |  2007| 2005| -0.0045766| 0.0157178|
#> |  2007| 2006| -0.0284475| 0.0181809|
#> |  2007| 2007| -0.0287814| 0.0162390|
#> 
#> 
#> P-value for pre-test of DID assumption:  0.23267
#> 

# without covariates
out2 <- att_gt(yname="lemp",
               tname="year",
               idname="countyreal",
               first.treat.name="first.treat",
               xformla=NULL,
               data=mpdta,
               printdetails=FALSE)
summary(out2)

#> 
#> Reference: Callaway, Brantly and Sant'Anna, Pedro.  "Difference-in-Differences with Multiple Time Periods." Working Paper <https://ssrn.com/abstract=3148250>, 2019. 
#> 
#> 
#> 
#> | group| time|        att|        se|
#> |-----:|----:|----------:|---------:|
#> |  2004| 2004| -0.0105032| 0.0232510|
#> |  2004| 2005| -0.0704232| 0.0309848|
#> |  2004| 2006| -0.1372587| 0.0364357|
#> |  2004| 2007| -0.1008114| 0.0343592|
#> |  2006| 2004|  0.0065201| 0.0233268|
#> |  2006| 2005| -0.0027508| 0.0195586|
#> |  2006| 2006| -0.0045946| 0.0177552|
#> |  2006| 2007| -0.0412245| 0.0202292|
#> |  2007| 2004|  0.0305067| 0.0150336|
#> |  2007| 2005| -0.0027259| 0.0163958|
#> |  2007| 2006| -0.0310871| 0.0178775|
#> |  2007| 2007| -0.0260544| 0.0166554|
#> 
#> 
#> P-value for pre-test of DID assumption:  0.16812
#>