ggdid

Function to plot MP objects

ggdid(mpobj, ylim = NULL, xlab = NULL, ylab = NULL, title = "Group",
  xgap = 1, ncol = 1)

Arguments

mpobj
an MP object

ylim
optional y limits for the plot; settng here makes the y limits the same across different plots

xlab
optional x-axis label

ylab
optional y-axis label

title
optional plot title

xgap
optional gap between the labels on the x-axis. For example, xgap=3 indicates that the labels should show up for every third value on the x-axis. The default is 1.

ncol
The number of columns to include in the resulting plot. The default is 1.