Multiplier Bootstrap

A function to take an influence function and use the multiplier bootstrap to compute standard errors and critical values for uniform confidence bands.

mboot(inf.func, DIDparams, pl = FALSE, cores = 1)

Arguments

inf.func

an influence function

DIDparams

DIDparams object

pl

whether or not to use parallel processing in the multiplier bootstrap, default=FALSE

cores

the number of cores to use with parallel processing, default=1

Value

list with elements

bres

results from each bootstrap iteration

V

variance matrix

se

standard errors

crit.val

a critical value for computing uniform confidence bands